# packages/contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/contributors/admin": {
      "get": {
        "operationId": "ContributorsController_listAdmin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContributorAdminDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Требуется авторизация"
          },
          "403": {
            "description": "Недостаточно прав"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "contributors"
        ]
      }
    },
    "/api/contributors": {
      "post": {
        "operationId": "ContributorsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContributorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributorAdminDto"
                }
              }
            }
          },
          "401": {
            "description": "Требуется авторизация"
          },
          "403": {
            "description": "Недостаточно прав"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "contributors"
        ]
      }
    },
    "/api/contributors/{id}": {
      "patch": {
        "operationId": "ContributorsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContributorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributorAdminDto"
                }
              }
            }
          },
          "401": {
            "description": "Требуется авторизация"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "404": {
            "description": "Участник не найден"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "contributors"
        ]
      },
      "delete": {
        "operationId": "ContributorsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Требуется авторизация"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "404": {
            "description": "Участник не найден"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "contributors"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portfolio API",
    "description": "Console Portfolio backend. The OpenAPI document is the contract: frontend types are generated from it.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "LocalizedTextDto": {
        "type": "object",
        "properties": {
          "ru": {
            "type": "string"
          },
          "en": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "ru"
        ]
      },
...
      "LocalizedTextInput": {
        "type": "object",
        "properties": {
          "ru": {
            "type": "string"
          },
          "en": {
            "type": "string"
          }
        },
        "required": [
          "ru"
        ]
      },
...
      "ContributorAdminDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "link": {
            "type": "string",
            "nullable": true
          },
          "order": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "image",
          "color",
          "link",
          "order"
        ]
      },
      "CreateContributorDto": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/LocalizedTextInput"
          },
          "image": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateContributorDto": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/LocalizedTextInput"
          },
          "image": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        }
      },
...
    }
  }
}